requirements.txt
Django==3.1.6
django-allauth==0.44.0
Pillow==8.1.0